Reverting away from anything in Loop
Was somewhat convoluted feeling, and digits were not always played.
I changed to using talkie pauses in place of using delay. 
I converted the battery level wavs to talkie hex for the female voice to match the digits when no wav file is available.
Edited the multiple wavs into one that contains pauses in the audio to allow the digits to play through, so if that wav so if it IS available, it plays with the talkie digits.
It works.
But again....with edit mode coming, and being able to use wavs for the digits, I just assume do plain digits here if you prefer, and lose the wav. I was just making it work for now. Unless you want to point me to where I would do that call to wav-not-talkie digits?