Add match blade effect to preon sounds
I would do this in my prop, but you had mentioned for the other preon matching thing that more universal might be better, so here it is.
Well, this works. It's on a #define so it's optional. If there's no EFFECT_PREONx in the blade style, standard EFFECT_PREON gets used.
If more than 3 preon.wavs exist, they use the current standard.
However I'm not sure we want to add additional effects while unused USER ones already exist, but it seems intuitive for a blade style to use the word PREON as opposed to USER2  ¯\_(ツ)_/¯.
I'm sure there's a more efficient way for all of this, and one that allows for more than 3 custom preons or even any effect, but this is a start I came up with for now. Maybe 3 is even fine and is enough? 
I'll update #216 to reflect this PR.