Description and added BLASTER_ENABLE_AUTO.
Added the copyright as a header, changed all the information on the header's comment to improve legibility. This will be important for future changes.
Also added BLASTER_ENABLE_AUTO and deprecated ENABLE_BLASTER_AUTO. Will add consistency to the next batch of defines added. And deprecated means that the usage will be discouraged but will still be 100% backwards compatible.