Update saber_sabersense_buttons.h
More redundant lines removed and tested. Since I'm not including Battle Mode, I've removed all references to it. Likewise the Sabersense defines that were listed near the top of the config - it turns out they didn't need to be there - they still work just appearing in the relevant places in the code. 
As such, hopefully it's a little tidier now. :)