Update saber_sabersense_buttons.h
I fear the indenting thing may be the thing that defeats this endeavour. I've tried running the code through a couple of different online formatters, but they all seem to do something slightly different with it, and I can't tell which is correct. They all seem also to do things that look wrong to me, but given that my formatting is wrong, maybe they are all correct, even with their respective differences. I've tried to read up on it and tried comparing with other ProffieOS prop files, but I just can't seem to find definitive correct methods. Anyway I've posted a version that, according to the reformatter, is correct. I've also done the bits I CAN do, like losing the TM and sorting double line spaces. I still need to sort out HAVE_EFFECT_QUOTE as I was kind of under the impression that the only file I could change was the prop file so that I wouldn't mess with anything else in the OS. But I'll have a crack at making that work. Although of course if the indenting does indeed prove to be the final impasse, everything else is pretty much irrelevant.