Sabersense Prop OS-7 Compatibility.
I've revisited OS-7 compatibility and, following your suggestion, used a better way of achieving it that requires no user input such as defines - it just works as is with either OS.
To differentiate systems, I've used the "Modes" define. Since modes are new to OS-8 but don't exist in OS-7, this is what tells the prop which system it's in.
I've also tweaked the Colour Change and Blaster Block code so that exiting the colour menu in OS-7 can be done with a short click of Power, the same as in OS-8.