Update saber_sabersense_buttons.h
OK, I'm hoping I've fixed the edge case problems with this. I've set it so that with big skips (5 or 10 presets at a time) it caps at the ends instead of wraps - I think it would be odd for a user to skip forward from preset 33 on a 36 preset system and find themselves back at preset 7.
I've hopefully also sorted the save/font file thing.