Add User Definable "Home" Preset.h
Intended for user who may have several sabers and they want all the same fonts on all sabers to be in the same order, so that preset navigation is the same for each hilt. BUt they still want a 'default' font to be the appropriate font for that hilt.
This feature makes that possible by letting the user define the home preset which replaces the skip-to-last preset. (Skip-to-last is easiest to lose because skip to first takes you one preset away from last, should you want it, due to wrapping).