Add Two-Way Array Cycling.
This has been touched on before, but always abandoned because it requires a define telling the system how many arrays are present. So this time I've wrapped the feature in a define so that it's there if people want it but doesn't place extra demands if they don't. I've also included a couple of safeguards to prevent invalid NUM_ARRAY values.

The only thing not covered is if people define fewer arrays than are present. But that's a fairly simple thing to trace in the config once they notice that behaviour does not match expectation.

I've also tweaked the UP/DOWN angles to make them less strict, and then changed the single-preset UP/DOWN forward/backward parameter to match.

Lastly Two-Way array cycling meant NextBladeArray was no longer accurate as it would apply to previous blade arrays too. Since all it really does is play the array idents, I changed it to be a more accurately descriptive IdentBladeArray.

Hope this all OK, and sorry to keep bombarding you with PRs - at least there is only a relatively small amount of code involved!

Thanks as always.
:)